fix: several fixes in windows forms app
frm commission fixes
frm courses fixes
frm student-courses fixes
frm my-courses fixes
frm academic-status fixes
frm qualify courses fixes

+++ Academia/UI.Desktop/Commission/frmCommission.cs
@@ -97,6 +97,7 @@ namespace UI.Desktop.Commission
 			{
 				try
 				{
+					if (Utilities.ConfirmDelete("esta comisión") != DialogResult.OK) return;
 					ApplicationCore.Model.Commission selectedCommission = (ApplicationCore.Model.Commission)lstvCommission.SelectedItems[0].Tag;
 					await commissionService.DeleteAsync(selectedCommission.Id);
 					LoadCommissions();
